# Swagger/UCSCGenomeAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Genome Data API",
    "description": "API for accessing genomic data from UCSC Genome Browser.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://api.genome.ucsc.edu/",
      "description": "Primary API Server"
    }
  ],
  "paths": {
    "/getData/sequence": {
      "get": {
        "tags": [
          "getData"
        ],
        "parameters": [
          {
            "name": "genome",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hubUrl",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chrom",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "end",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/getData/track": {
      "get": {
        "tags": [
          "getData"
        ],
        "parameters": [
          {
            "name": "genome",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "track",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hubUrl",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chrom",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "end",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "jsonOutputArrays",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "maxItemsOutput",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/list/publicHubs": {
      "get": {
        "tags": [
          "list"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/list/ucscGenomes": {
      "get": {
        "tags": [
          "list"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/list/hubGenomes": {
      "get": {
        "tags": [
          "list"
        ],
        "parameters": [
          {
            "name": "hubUrl",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/list/files": {
      "get": {
        "tags": [
          "list"
        ],
        "parameters": [
          {
            "name": "genome",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "style": "form",
            "schema": {
              "$ref": "#/components/schemas/Format"
            }
          },
          {
            "name": "maxItemsOutput",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/list/tracks": {
      "get": {
        "tags": [
          "list"
        ],
        "parameters": [
          {
            "name": "genome",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hubUrl",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trackLeavesOnly",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/list/chromosomes": {
      "get": {
        "tags": [
          "list"
        ],
        "parameters": [
          {
            "name": "genome",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hubUrl",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "track",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/list/schema": {
      "get": {
        "tags": [
          "list"
        ],
        "parameters": [
          {
            "name": "genome",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hubUrl",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "track",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/search": {
      "get": {
        "tags": [
          "search"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "genome",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categories",
            "in": "query",
            "style": "form",
            "schema": {
              "$ref": "#/components/schemas/Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Category": {
        "enum": [
          "helpDocs",
          "publicHubs",
          "trackDb"
        ],
        "type": "string"
      },
      "Format": {
        "enum": [
          "text"
        ],
        "type": "string"
      }
    }
  }
}